+++ src/webview-ui/components/propertyItem.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { VSCodeStyleProperty } from '../hooks/useVSCodeStyles';
 import { useCodeTheme } from '../providers/codeThemeProvider';
+import { PropertyColor } from './propertyColor';
 
 export interface IPropertyItemProps {
   property: VSCodeStyleProperty;
@@ -13,20 +14,23 @@ export const PropertyItem: React.FunctionComponent<IPropertyItemProps> = ({ prop
 
   if (value.startsWith("#") || value.startsWith("rgb")) {
     return (
-      <div className={`grid grid-cols-2 col`}>
-        <div className={`flex flex-col py-2 justify-center`}>
-          <div><strong>CSS Variable</strong>: {cssVariable}</div>
-          <div className={`flex items-center`}><strong>Value</strong>: <div className="inline-block h-5 w-5 m-2 border border-black" style={{ backgroundColor: value }}></div> {value}</div>
-        </div>
+      <div className={`grid grid-cols-2`}>
+       <PropertyColor 
+        property={property}
+        value={value}
+        cssVariable={cssVariable}
+        codeNotition={codeNotition} />
 
         {
           compareProperty ? (
-            <div className={`flex flex-col py-2 pl-2 border-l border-gray-200 bg-gray-50`}>
-              <div><strong>CSS Variable</strong>: {cssVariable}</div>
-              <div className={`flex items-center`}><strong>Value</strong>: <div className="inline-block h-5 w-5 m-2 border border-black" style={{ backgroundColor: compareProperty }}></div> {compareProperty}</div>
-            </div>
+            <PropertyColor 
+             property={property}
+             value={compareProperty}
+             cssVariable={cssVariable}
+             codeNotition={codeNotition}
+             isSecondary />
           ) : (
-            name && ( <div className='bg-gray-50 border-l border-gray-200 block'></div> )
+            name && ( <div className='bg-gray-50 block'></div> )
           )
         }
       </div>